fix(condo): moved errors back and removed function errors
